Change safe_malloc's size parameter from (unsigned int) to size_t.

+++ lib.c
@@ -281,7 +281,7 @@ void *safe_calloc (size_t nmemb, size_t size)
   return p;
 }
 
-void *safe_malloc (unsigned int siz)
+void *safe_malloc (size_t siz)
 {
   void *p;
